+++ src/constants/contentTypes.js
@@ -1,13 +1,13 @@
 const contentTypes = {
-  AGGREGATED: 'aggregated',
-  ABSTRACT: 'abstract',
-  EBOOK: 'ebook',
-  EJOURNAL: 'ejournal',
-  MIXED_CONTENT: 'mixedContent',
-  ONLINE_REFERENCE: 'onlineReference',
-  PRINT: 'print',
-  STREAMING_MEDIA: 'streamingMedia',
-  UNKNOWN: 'unknown',
+  AGGREGATED: 'Aggregated Full Text',
+  ABSTRACT: 'Abstract and Index',
+  EBOOK: 'E-Book',
+  EJOURNAL: 'E-Journal',
+  MIXED_CONTENT: 'Mixed Content',
+  ONLINE_REFERENCE: 'Online Reference',
+  PRINT: 'Print',
+  STREAMING_MEDIA: 'Streaming Media',
+  UNKNOWN: 'Unknown',
 };
 
 export default contentTypes;