+++ src/constants/filterConfigs.js
@@ -66,18 +66,18 @@ export const publicationTypeFilterConfig = {
   defaultValue: 'all',
   options: [
     { label: <FormattedMessage id="ui-eholdings.filter.all" />, value: 'all' },
-    { label: <FormattedMessage id="ui-eholdings.filter.pubType.audioBook" />, value: 'audiobook' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.audiobook" />, value: 'audiobook' },
     { label: <FormattedMessage id="ui-eholdings.filter.pubType.book" />, value: 'book' },
-    { label: <FormattedMessage id="ui-eholdings.filter.pubType.bookSeries" />, value: 'bookseries' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.book_series" />, value: 'bookseries' },
     { label: <FormattedMessage id="ui-eholdings.filter.pubType.database" />, value: 'database' },
     { label: <FormattedMessage id="ui-eholdings.filter.pubType.journal" />, value: 'journal' },
     { label: <FormattedMessage id="ui-eholdings.filter.pubType.newsletter" />, value: 'newsletter' },
     { label: <FormattedMessage id="ui-eholdings.filter.pubType.newspaper" />, value: 'newspaper' },
     { label: <FormattedMessage id="ui-eholdings.filter.pubType.proceedings" />, value: 'proceedings' },
     { label: <FormattedMessage id="ui-eholdings.filter.pubType.report" />, value: 'report' },
-    { label: <FormattedMessage id="ui-eholdings.filter.pubType.streamingAudio" />, value: 'streamingaudio' },
-    { label: <FormattedMessage id="ui-eholdings.filter.pubType.streamingVideo" />, value: 'streamingvideo' },
-    { label: <FormattedMessage id="ui-eholdings.filter.pubType.thesisdissertation" />, value: 'thesisdissertation' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.streaming_audio" />, value: 'streamingaudio' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.streaming_video" />, value: 'streamingvideo' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.thesis_dissertation" />, value: 'thesisdissertation' },
     { label: <FormattedMessage id="ui-eholdings.filter.pubType.website" />, value: 'website' },
     { label: <FormattedMessage id="ui-eholdings.filter.pubType.unspecified" />, value: 'unspecified' }
   ]
@@ -93,10 +93,10 @@ export const contentTypeFilterConfig = {
     { label: <FormattedMessage id="ui-eholdings.filter.contentType.abstract" />, value: 'abstractandindex' },
     { label: <FormattedMessage id="ui-eholdings.filter.contentType.ebook" />, value: 'ebook' },
     { label: <FormattedMessage id="ui-eholdings.filter.contentType.ejournal" />, value: 'ejournal' },
-    { label: <FormattedMessage id="ui-eholdings.filter.contentType.mixedContent" />, value: 'mixedcontent' },
-    { label: <FormattedMessage id="ui-eholdings.filter.contentType.onlineReference" />, value: 'onlinereference' },
+    { label: <FormattedMessage id="ui-eholdings.filter.contentType.mixed_content" />, value: 'mixedcontent' },
+    { label: <FormattedMessage id="ui-eholdings.filter.contentType.online_reference" />, value: 'onlinereference' },
     { label: <FormattedMessage id="ui-eholdings.filter.contentType.print" />, value: 'print' },
-    { label: <FormattedMessage id="ui-eholdings.filter.contentType.streamingMedia" />, value: 'streamingmedia' },
+    { label: <FormattedMessage id="ui-eholdings.filter.contentType.streaming_media" />, value: 'streamingmedia' },
     { label: <FormattedMessage id="ui-eholdings.filter.contentType.unknown" />, value: 'unknown' }
   ]
 };